+++ CHANGELOG.md
@@ -5,11 +5,35 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.0.1-alpha] - 2023-03-01
+## [v0.0.2] - 2023-03-20
+
+### Added
+
+- Config
+  - Changed .yml to .yaml
+  - Added more config settings
+- Paid Relay
+  - Core lightning supported
+- Command line tool
+  - whitelist
+    - list
+    - add
+    - remove
+  - invoice
+    - create
+    - check
+  - port
+  - loglevel
+
+### Fixed
+- Project Structure
+
+## [v0.0.1] - 2023-03-01
 
 ### Added
 
 - Initial alpha version of nostr citadel.
 
-[unreleased]: https://github.com/mrjohnsson77/nostr-citadel/compare/v0.0.1...HEAD
-[0.0.1-alpha]: https://github.com/mrjohnsson77/nostr-citadel/releases/tag/v0.0.1
+[unreleased]: https://github.com/mrjohnsson77/nostr-citadel/compare/v0.0.2...HEAD
+[v0.0.1]: https://github.com/mrjohnsson77/nostr-citadel/releases/tag/v0.0.1
+[v0.0.2]: https://github.com/mrjohnsson77/nostr-citadel/releases/tag/v0.0.2